Move F7-specific stuff to F7 implementation library
