+++ source/Meadow.Core/NativeNetworkAdapterCollection.cs
@@ -61,9 +61,6 @@ namespace Meadow.Devices
                         case NetworkInterfaceType.Wireless80211:
                             _adapters.Add(new WiFiNetworkAdapter(ni));
                             break;
-                        case NetworkInterfaceType.Ppp:
-                            _adapters.Add(new CellNetworkAdapter(ni));
-                            break;
                     }
                 }
             }
